Add admin privileges to review and comment deletion

+++ backend/src/controllers/comment.controller.js
@@ -98,8 +98,8 @@ async function update(req, res) {
 
 async function deleteComment(req, res) {
   try {
-    const { user_id } = req.user;
-    const { comment_id } = req.body;
+    const { user_id: requestingUser } = req.user;
+    const { user_id, comment_id } = req.body;
 
     // Find if comment exists
     const { rows: commented } = await pool.query(
@@ -111,10 +111,23 @@ async function deleteComment(req, res) {
       return res.status(404).json({ error: "Review not found" });
     }
 
-    await pool.query(
-      `DELETE FROM comments WHERE user_id=$1 AND comment_id=$2`,
-      [user_id, comment_id]
+    const { rows: user } = await pool.query(
+      `
+      SELECT * from users WHERE user_id=$1
+      `,
+      [requestingUser]
     );
+
+    // Delete comment if user_id matches or requesting user is admin
+    if (
+      user.length !== 0 &&
+      (Number(user[0].user_id) === Number(user_id) || user[0].admin)
+    ) {
+      await pool.query(
+        `DELETE FROM comments WHERE user_id=$1 AND comment_id=$2`,
+        [user_id, comment_id]
+      );
+    }
 
     res.status(200).end();
   } catch (e) {
